[refactor] - 필터링 로직을 zustand store에서 useFilter hook으로 변경
